+++ app/controllers/carto/api/public/oauth_apps_controller.rb
@@ -12,6 +12,8 @@ module Carto
 
         before_action :load_user
         before_action :load_owned_app, only: [:show, :update, :regenerate_secret, :destroy]
+        before_action :load_granted_app, only: :revoke
+        before_action :load_index_params, only: [:index, :index_granted]
         before_action :engine_required
 
         setup_default_rescues
@@ -19,21 +21,13 @@ module Carto
         VALID_ORDER_PARAMS = [:name, :updated_at, :restricted, :user_id].freeze
 
         def index
-          page, per_page, order = page_per_page_order_params(VALID_ORDER_PARAMS)
           oauth_apps = user_or_organization_apps
-          filtered_oauth_apps = Carto::PagedModel.paged_association(oauth_apps, page, per_page, order)
-          result = filtered_oauth_apps.map { |oauth_app| OauthAppPresenter.new(oauth_app).to_hash }
+          render_paged(oauth_apps) { |params| api_v4_oauth_apps_url(params) }
+        end
 
-          render_jsonp(
-            paged_result(
-              result: result,
-              total_count: oauth_apps.size,
-              page: page,
-              per_page: per_page,
-              params: params.except('controller', 'action')
-            ) { |params| api_v4_oauth_apps_url(params) },
-            200
-          )
+        def index_granted
+          oauth_apps = @user.granted_oauth_apps
+          render_paged(oauth_apps) { |params| api_v4_oauth_apps_index_granted_url(params) }
         end
 
         def show
@@ -61,6 +55,12 @@ module Carto
           head :no_content
         end
 
+        def revoke
+          oauth_app_user = @oauth_app.oauth_app_users.where(user_id: @user.id).first
+          oauth_app_user.destroy!
+          head :no_content
+        end
+
         private
 
         def load_user
@@ -72,10 +72,23 @@ module Carto
           raise ActiveRecord::RecordNotFound.new unless owned?
         end
 
+        def load_granted_app
+          @oauth_app = Carto::OauthApp.find(params[:id])
+          raise ActiveRecord::RecordNotFound.new unless granted?
+        end
+
         def owned?
           return true if @oauth_app.user_id == @user.id
 
           @user.organization_admin? && @user.organization == @oauth_app.user.organization
+        end
+
+        def granted?
+          @user.granted_oauth_apps.include?(@oauth_app)
+        end
+
+        def load_index_params
+          @page, @per_page, @order = page_per_page_order_params(VALID_ORDER_PARAMS)
         end
 
         def user_or_organization_apps
@@ -89,6 +102,20 @@ module Carto
           params.permit(:name, :icon_url, redirect_uris: [])
         end
 
+        def render_paged(oauth_apps)
+          filtered_oauth_apps = Carto::PagedModel.paged_association(oauth_apps, @page, @per_page, @order)
+          result = filtered_oauth_apps.map { |oauth_app| OauthAppPresenter.new(oauth_app).to_hash }
+
+          enriched_response = paged_result(
+            result: result,
+            total_count: oauth_apps.size,
+            page: @page,
+            per_page: @per_page,
+            params: params.except('controller', 'action')
+          ) { |params| yield(params) }
+
+          render_jsonp(enriched_response, 200)
+        end
       end
     end
   end